+++ src/Persistence/Game.GameStateSource.cs
@@ -1,5 +1,6 @@
 using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
+using CivOne.Enums;
 using CivOne.Services.GlobalWarming;
 using CivOne.Tiles;
 using CivOne.Units;
@@ -20,6 +21,8 @@ namespace CivOne
 		public ushort PlayerFutureTech => HumanPlayer?.FutureTechCount ?? _playerFutureTech;
 
 		bool IGameSnapshotSource.DisableBuddyCivilizationRespawn => DisableBuddyCivilizationRespawn;
+
+		BarbarianActivity IGameSnapshotSource.BarbarianActivity => BarbarianActivity;
 
 		[SuppressMessage("Microsoft.Design", "CA1002:DoNotExposeGenericLists", Justification = "A list is suffice and changing it would require unnecessary changes to the GameState and related code.")]
 		public List<ReplayData> ReplayData => _replayData;